requirements.txt
antlr4-python3-runtime==4.8
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
flake8==3.8.4
hydra-core==1.0.3
idna==2.10
importlib-resources==3.0.0
isort==5.5.4
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
omegaconf==2.0.2
pycodestyle==2.6.0
pyflakes==2.2.0
PyYAML==5.3.1
requests==2.24.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
